V4L/DVB (6963): tda18271: store IF frequency in a u16 instead of u32
Store IF Frequency in a u16 instead of a u32.  Multiply by 1000 before use.

+++ drivers/media/dvb/frontends/tda18271.h
@@ -25,7 +25,7 @@
 #include "dvb_frontend.h"
 
 struct tda18271_std_map_item {
-	u32 if_freq;
+	u16 if_freq;
 	u8 std_bits;
 };
